refactor: remove Button/index
