--- site/src/components/Button/index.ts
@@ -1,3 +1,0 @@
-export * from "./CopyButton"
-export * from "./LoadingButton"
-export * from "./SplitButton"
