+++ site/src/components/CodeExample/CodeExample.tsx
@@ -1,7 +1,7 @@
 import { makeStyles } from "@material-ui/core/styles"
 import React from "react"
 import { MONOSPACE_FONT_FAMILY } from "../../theme/constants"
-import { CopyButton } from "../Button"
+import { CopyButton } from "../Button/CopyButton"
 
 export interface CodeExampleProps {
   code: string
