+++ site/src/components/index.tsx
@@ -1,3 +1,2 @@
-export * from "./Button"
 export * from "./EmptyState"
 export * from "./Page"
